redux succesfully removed
