+++ actions(not used)/index.js
@@ -1,0 +1,38 @@
+import * as types from '../constants/ActionTypes'
+
+//action creators
+export function addTodo(text) {
+  return { type: types.ADD_TODO, text }
+}
+
+export function deleteTodo(id) {
+  return { type: types.DELETE_TODO, id }
+}
+
+export function editTodo(id, text) {
+  return { type: types.EDIT_TODO, id, text }
+}
+
+export function completeTodo(id) {
+  return { type: types.COMPLETE_TODO, id }
+}
+
+export function completeAll() {
+  return { type: types.COMPLETE_ALL }
+}
+
+export function clearCompleted() {
+  return { type: types.CLEAR_COMPLETED }
+}
+
+export function receiveInitialTodos() {
+  return { type: types.RECEIVE_INITIAL_TODOS }
+}
+
+
+export function receiveInitialTodos(json) {
+  return {
+    type: RECEIVE_INITIAL_TODOS,
+    todos: json.data.children.map(child => child.data)
+  }
+}
