+++ constants(not used)/ActionTypes.js
@@ -1,0 +1,11 @@
+export const ADD_TODO = 'ADD_TODO'
+export const DELETE_TODO = 'DELETE_TODO'
+export const EDIT_TODO = 'EDIT_TODO'
+export const COMPLETE_TODO = 'COMPLETE_TODO'
+export const COMPLETE_ALL = 'COMPLETE_ALL'
+export const CLEAR_COMPLETED = 'CLEAR_COMPLETED'
+
+export const REQUEST_INITIAL_TODOS = 'REQUEST_INITIAL_TODOS'
+export const REQUEST_INITIAL_TODOS_FAILURE = 'REQUEST_INITIAL_TODOS_FAILURE'
+export const REQUEST_INITIAL_TODOS_SUCCESS = 'REQUEST_INITIAL_TODOS_SUCCESS'
+export const RECEIVE_INITIAL_TODOS = 'RECEIVE_INITIAL_TODOS'
