+++ constants(not used)/TodoFilters.js
@@ -1,0 +1,3 @@
+export const SHOW_ALL = 'show_all'
+export const SHOW_COMPLETED = 'show_completed'
+export const SHOW_ACTIVE = 'show_active'
