+++ containers(not used)/App.js
@@ -1,0 +1,47 @@
+import React, { Component, PropTypes } from 'react'
+import { bindActionCreators } from 'redux'
+import { connect } from 'react-redux'
+import Header from '../components/Header'
+import MainSection from '../components/MainSection'
+import * as TodoActions from '../actions'
+import CurrentDay from '../components/CurrentDay'
+import ShowCalendarBtn from '../components/ShowCalendarBtn'
+import InitialTodos from '../components/InitialTodos'
+
+
+class App extends Component {
+  render() {
+    const { todos, actions } = this.props
+    return (
+      <div>
+        <CurrentDay />
+        <ShowCalendarBtn />
+        <Header addTodo={actions.addTodo} />
+        <MainSection todos={todos} actions={actions} />
+        
+      </div>
+    )
+  }
+}
+
+App.propTypes = {
+  todos: PropTypes.array.isRequired,
+  actions: PropTypes.object.isRequired
+}
+
+function mapStateToProps(state) {
+  return {
+    todos: state.todos
+  }
+}
+
+function mapDispatchToProps(dispatch) {
+  return {
+    actions: bindActionCreators(TodoActions, dispatch)
+  }
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(App)
