+++ package.json
@@ -26,6 +26,7 @@
     "compression": "^1.6.1",
     "express": "^4.14.0",
     "if-env": "^1.0.0",
+    "jquery": "^3.1.0",
     "knex": "^0.11.7",
     "moment": "^2.14.1",
     "pg": "^6.0.2",
@@ -35,7 +36,8 @@
     "react-dom": "^0.14.7",
     "react-redux": "^4.2.1",
     "react-router": "^2.5.2",
-    "redux": "^3.2.1"
+    "redux": "^3.2.1",
+    "sqlite3": "^3.1.4"
   },
   "devDependencies": {
     "babel-core": "^6.5.1",
