+++ package.json(not used)
@@ -1,0 +1,66 @@
+{
+  "name": "daily",
+  "version": "0.0.1",
+  "description": "Todo app based on daily tasks for personal useage",
+  "scripts": {
+    "start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
+    "start:dev": "webpack-dev-server --inline --content-base public/ --history-api-fallback",
+    "start:prod": "webpack && node --debug ./server/server.js",
+    "test": "cross-env NODE_ENV=test mocha --recursive --compilers js:babel-register --require ./test/setup.js",
+    "test:watch": "npm test -- --watch"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/navyboys/daily.git"
+  },
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/navyboys/daily/issues"
+  },
+  "homepage": "https://daily.io",
+  "dependencies": {
+    "babel-polyfill": "^6.3.14",
+    "body-parser": "^1.15.2",
+    "bookshelf": "^0.10.0",
+    "classnames": "^2.1.2",
+    "compression": "^1.6.1",
+    "express": "^4.14.0",
+    "if-env": "^1.0.0",
+    "knex": "^0.11.7",
+    "moment": "^2.14.1",
+    "pg": "^6.0.2",
+    "react": "^0.14.7",
+    "react-bootstrap": "^0.29.5",
+    "react-datetime": "^2.3.2",
+    "react-dom": "^0.14.7",
+    "react-redux": "^4.2.1",
+    "react-router": "^2.5.2",
+    "redux": "^3.2.1",
+    "sqlite3": "^3.1.4"
+  },
+  "devDependencies": {
+    "babel-core": "^6.5.1",
+    "babel-loader": "^6.2.2",
+    "babel-preset-es2015": "^6.5.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-react-hmre": "^1.1.1",
+    "babel-register": "^6.3.13",
+    "cross-env": "^1.0.7",
+    "enzyme": "^2.2.0",
+    "expect": "^1.8.0",
+    "http-server": "^0.8.5",
+    "jsdom": "^5.6.1",
+    "mocha": "^2.2.5",
+    "node-libs-browser": "^0.5.2",
+    "raw-loader": "^0.5.1",
+    "react-addons-test-utils": "^0.14.7",
+    "sqlite3": "^3.1.4",
+    "strftime": "^0.9.2",
+    "style-loader": "^0.12.3",
+    "todomvc-app-css": "^2.0.1",
+    "webpack": "^1.9.11",
+    "webpack-dev-middleware": "^1.2.0",
+    "webpack-dev-server": "^1.14.1",
+    "webpack-hot-middleware": "^2.9.1"
+  }
+}
