+++ public/index.html
@@ -20,7 +20,7 @@
     <link type="text/css" rel="stylesheet" href="./stylesheet.css" />
   </head>
   <body>
-    <div class="todoapp" id="root">
+    <div class="todoapp" id="main">
     </div>
     <script src="/static/bundle.js"></script>
   </body>
