+++ public/stylesheet.css
@@ -2,21 +2,19 @@
  * https://github.com/arqex/react-datetime
  */
 
-.main {
-  background-color: green;
-}
-
-.todoList {
-  background-color: red;
-}
+/*.todoList {
+  position: absolute;
+  margin-top: 200px
+}*/
 
 /*.pageHeader {
   background-color: red;
 }*/
 
-.header {
-  background-color: #90C5A9
-}
+/*div.header {
+  position: relative;
+
+}*/
 
 .rdt {
 	position: relative;
@@ -31,7 +29,7 @@
 	position: absolute;
   width: 250px;
   padding: 4px;
-  margin-top: 1px;
+  margin-top: 0px;
   z-index: 99999 !important;
   background: #fff;
   box-shadow: 0 1px 3px rgba(0,0,0,.1);
@@ -237,7 +235,7 @@ td.rdtYear:hover {
 	margin-top: 43px;
 }
 
-/*Below is the todoMVC app css*/
+
 html,
 body {
 	margin: 0;
@@ -310,22 +308,50 @@ body {
 }
 .showCalendar {
   position: absolute;
-  top: -15px;
-  right: 0px;
+  top: -27px;
+  right: 15px;
 }
 .calendar {
   position: absolute;
-  top: -23px;
+  top: -35px;
   margin-left: auto;
   margin-right: auto;
   width: 100%;
   color: #EF6D3B
 }
+.table_rows {
+  font-size: 25px;
+
+}
+
+.table_cells {
+  position: relative;;
+  /*top: -50px;*/
+  /*margin-right: 50px*/
+  right: -80px;
+  font-size: 30px
+}
+
+.markComplete {
+  position: relative;
+	top: 0px;
+	left: 475px;
+	bottom: 0;
+	width: 30px;
+	height: 30px;
+	margin-top: 0;
+	font-size: 20px;
+	/*color: #cc9a9a;*/
+  color: #90C5A9;
+	margin-bottom: 0px;
+	transition: color 0.2s ease-out;
+}
+
 .todoapp h1 {
 	position: absolute;
-	top: -155px;
+	top: -153px;
 	width: 100%;
-	font-size: 100px;
+	font-size: 80px;
 	font-weight: 100;
 	text-align: center;
 	/*color: rgba(175, 47, 47, 0.15);*/
@@ -439,20 +465,6 @@ label[for='toggle-all'] {
 	appearance: none;
 }
 
-.todo-list li .markComplete {
-  position: absolute;
-	top: 0;
-	left: 10px;
-	bottom: 0;
-	width: 30px;
-	height: 30px;
-	margin: auto 0;
-	font-size: 20px;
-	/*color: #cc9a9a;*/
-  color: #90C5A9;
-	margin-bottom: 15px;
-	transition: color 0.2s ease-out;
-}
 
 .todo-list li .toggle:checked:after {
 	content: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="40" height="40" viewBox="-10 -18 100 135"><circle cx="50" cy="50" r="50" fill="none" stroke="#bddad5" stroke-width="3"/><path fill="#5dc2af" d="M72 25L42 71 27 56l-4 4 20 20 34-52z"/></svg>');
@@ -471,18 +483,27 @@ label[for='toggle-all'] {
 	color: #d9d9d9;
 	text-decoration: line-through;
 }
+.completed {
+  color: #d9d9d9;
+  text-decoration: line-through;
+}
+.removed {
+  display: none
+}
 
-.todo-list li .destroy {
-	position: absolute;
-	top: 0;
-	right: 10px;
+
+.destroy {
+  position: relative;
+	top: 0px;
+	left: 15px;
 	bottom: 0;
 	width: 30px;
 	height: 30px;
-	margin: auto 0;
+	margin-top: 0;
 	font-size: 20px;
+	/*color: #cc9a9a;*/
   color: #90C5A9;
-	margin-bottom: 15px;
+	margin-bottom: 0px;
 	transition: color 0.2s ease-out;
 }
