+++ reducers(not used)/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux'
+import todos from './todos'
+
+const rootReducer = combineReducers({
+  todos
+})
+
+export default rootReducer
