--- reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux'
-import todos from './todos'
-
-const rootReducer = combineReducers({
-  todos
-})
-
-export default rootReducer
