+++ webpack.config.js
@@ -5,7 +5,7 @@ module.exports = {
   devtool: 'cheap-module-eval-source-map',
   entry: [
     'webpack-hot-middleware/client',
-    './index'
+    './src/app'
   ],
   output: {
     path: path.join(__dirname, 'dist'),
